+++ aims_data_platform/__init__.py
@@ -33,7 +33,7 @@ from dq_framework import (
     FileSystemHandler,
 )
 
-__version__ = "1.4.0"
+__version__ = "1.3.1"
 __all__ = [
     "config",
     "Config",